+++ Project/src/DirectionalLight.h
@@ -7,7 +7,7 @@ public:
     DirectionalLight(glm::vec3 colour, GLfloat ambientIntensity, GLfloat diffuseIntensity, glm::vec3 direction );
     ~DirectionalLight();
     
-    void Use(GLfloat ambientIntensityLocation, GLfloat colourLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+    void Use(GLuint ambientIntensityLocation, GLuint colourLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
     
 private:
     glm::vec3 direction;
